+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.0-dev.2]
+Create a separate ReadonlyDataSource abstraction, so that data sources
+that only allow querying would not have to implement modifiable 
+DataSource class.
+
 ## [1.0.0-dev.1]
 Initial implementation of collection oriented repositories with no
 actual implementations at the moment.